requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.2
django-admin-interface==0.26.0
django-colorfield==0.13.0
django-constance==4.3.2
django-debug-toolbar==5.2.0
django-jazzmin==3.0.1
django-simple-history==3.8.0
django-unfold==0.55.3
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
Unidecode==1.3.8
